requirements.txt
Flask==3.0.0
requests==2.31.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
werkzeug==3.0.1

